+++ tasks/commons/erb.rb
@@ -1,0 +1,5 @@
+class ErbFromOStruct < OpenStruct
+  def render(template)
+    ERB.new(template).result(binding)
+  end
+end